kvs_backend/requirements.txt
fastapi==0.101.1
uvicorn==0.23.2
pydantic==2.1.1
python-dotenv==1.0.0
azure-identity==1.13.0
azure-keyvault-secrets==4.7.0
azure-keyvault-certificates==4.7.0
azure-mgmt-costmanagement==3.0.0
azure-mgmt-resource==23.0.1
azure-mgmt-keyvault
python-dateutil==2.8.2
tensorflow==2.16.1
azure-data-tables
pydantic[email]
